+++ src/lib.rs
@@ -13,3 +13,4 @@ mod test;
 mod test_util;
 
 pub mod cli;
+extern crate log;