+++ src/program.rs
@@ -7,7 +7,7 @@ use polonius_parser::{
     parse_input,
 };
 
-use crate::facts::{AllFacts, Loan, Point, Region};
+use crate::facts::{AllFacts, Loan, Point, Region, Variable};
 use crate::intern::InternerTables;
 
 /// A structure to hold and deduplicate facts
@@ -20,6 +20,11 @@ struct Facts {
     outlives: BTreeSet<(Region, Region, Point)>,
     region_live_at: BTreeSet<(Region, Point)>,
     invalidates: BTreeSet<(Point, Loan)>,
+    var_defined: BTreeSet<(Variable, Point)>,
+    var_used: BTreeSet<(Variable, Point)>,
+    var_drop_used: BTreeSet<(Variable, Point)>,
+    var_uses_region: BTreeSet<(Variable, Region)>,
+    var_drops_region: BTreeSet<(Variable, Region)>,
 }
 
 impl From<Facts> for AllFacts {
@@ -32,6 +37,11 @@ impl From<Facts> for AllFacts {
             outlives: facts.outlives.into_iter().collect(),
             region_live_at: facts.region_live_at.into_iter().collect(),
             invalidates: facts.invalidates.into_iter().collect(),
+            var_defined: facts.var_defined.into_iter().collect(),
+            var_used: facts.var_used.into_iter().collect(),
+            var_drop_used: facts.var_drop_used.into_iter().collect(),
+            var_uses_region: facts.var_uses_region.into_iter().collect(),
+            var_drops_region: facts.var_drops_region.into_iter().collect(),
         }
     }
 }
@@ -52,6 +62,24 @@ pub(crate) fn parse_from_program(
             .iter()
             .map(|region| tables.regions.intern(region)),
     );
+
+    facts
+        .var_drops_region
+        .extend(input.var_drops_region.iter().map(|(variable, region)| {
+            (
+                tables.variables.intern(variable),
+                tables.regions.intern(region),
+            )
+        }));
+
+    facts
+        .var_uses_region
+        .extend(input.var_uses_region.iter().map(|(variable, region)| {
+            (
+                tables.variables.intern(variable),
+                tables.regions.intern(region),
+            )
+        }));
 
     for block in &input.blocks {
         let block_name = &block.name;
@@ -187,6 +215,20 @@ fn emit_fact(facts: &mut Facts, fact: &Fact, point: Point, tables: &mut Interner
             // region_live_at: a region can be manually set live on both Start and Mid points
             // but will mostly be computed and emitted automatically
             facts.region_live_at.insert((region, point));
+        }
+
+        // facts: var_defined(V, P)
+        Fact::DefineVariable { ref variable } => {
+            // var_defined: a variable is overwritten here
+            let variable = tables.variables.intern(variable);
+            facts.var_defined.insert((variable, point));
+        }
+
+        // facts: var_used(V, P)
+        Fact::UseVariable { ref variable } => {
+            // var_used: a variable is used here
+            let variable = tables.variables.intern(variable);
+            facts.var_used.insert((variable, point));
         }
     };
 }